API v0.1.1 Complete, properties for web-server

+++ api/src/main/java/naakcii/by/api/service/CategoryService.java
@@ -8,11 +8,11 @@ public interface CategoryService {
 
     CategoryDTO create(CategoryDTO categoryDTO);
 
-    CategoryDTO delete(int id);
+    CategoryDTO delete(Long id);
 
     List<CategoryDTO> findAll();
 
-    CategoryDTO findById(int id);
+    CategoryDTO findById(Long id);
 
     CategoryDTO update(CategoryDTO categoryDTO);
 }
